# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.BookingCreatable:
    properties:
      includeTableCost:
        type: boolean
      note:
        type: string
      productsId:
        items:
          type: integer
        type: array
      quantities:
        items:
          type: integer
        type: array
      reason:
        type: string
      secretCode:
        type: integer
      tableId:
        type: integer
    type: object
...
  models.response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      statusCode:
        type: integer
    type: object
host: localhost:3001
info:
  contact:
    email: dso.intern.xuanhoa@gmail.com
    name: Xuan Hoa Le
  description: Go Food Delivery CMS - Swagger API Discovery
  title: Go Food Delivery CMS - Swagger API Discovery
  version: "1.0"
paths:
...
  /booking:
    post:
      consumes:
      - application/json
      description: Create Booking
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Booking
        in: body
        name: booking
        required: true
        schema:
          $ref: '#/definitions/models.BookingCreatable'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: Create Booking
      tags:
      - booking
  /booking/:
    get:
      consumes:
      - application/json
      description: GetDetail Booking
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Order Id
        in: query
        name: orderId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: GetDetail Booking
      tags:
      - bills
  /booking/accepted:
    post:
      consumes:
      - application/json
      description: Accept Order
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Order Id
        in: query
        name: orderId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: Accept Order
      tags:
      - bills
  /booking/all:
    get:
      consumes:
      - application/json
      description: Get All Booking
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: Get All Booking
      tags:
      - bills
  /booking/employee:
    get:
      consumes:
      - application/json
      description: Get Top 10 Detail Booking By EmployeeId
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: Get Top 10 Detail Booking By EmployeeId
      tags:
      - bills
  /booking/finished:
    post:
      consumes:
      - application/json
      description: Finish Order
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Order Id
        in: query
        name: orderId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: Finish Order
      tags:
      - bills
  /booking/refund/{orderId}:
    patch:
      consumes:
      - application/json
      description: Refund Booking
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Table Id
        in: path
        name: tableId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: Refund Booking
      tags:
      - bills
  /booking/rejected:
    post:
      consumes:
      - application/json
      description: Reject Order
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Order Id
        in: query
        name: orderId
        required: true
        type: integer
      - description: Reason
        in: query
        name: reason
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: Reject Order
      tags:
      - bills
  /booking/table/{tableId}:
    get:
      consumes:
      - application/json
      description: Get Serve Booking By TableId
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Table Id
        in: path
        name: tableId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: Get Serve Booking By TableId
      tags:
      - bills
  /booking/table/preparing/{tableId}:
    get:
      consumes:
      - application/json
      description: Get Preparing Booking By Table Id
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Table Id
        in: path
        name: tableId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: Get Preparing Booking By Table Id
      tags:
      - bills
  /booking/table/rejected/{tableId}:
    get:
      consumes:
      - application/json
      description: Get Rejected Booking By TableId
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Table Id
        in: path
        name: tableId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: Get Rejected Booking By TableId
      tags:
      - bills
...
swagger: "2.0"